bankAccounts: refactoring
